refactor: only use one class for multi-modal and normal
